Keep API route helpers outside Next page exports
